requirements.txt
asgiref==3.11.0
django==6.0
django-cors-headers==4.9.0
django-dashub==3.0.7
django-filter==25.2
django-tinymce==5.0.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pillow==12.0.0
pyjwt==2.10.1
sqlparse==0.5.4
